remove mistaken comments
Change readme file to be more descriptive on how to get this app
running.  Need to do this since someone who downloads this code may not
have access to Github's built in wiki to see directions on how to get
this app running.

Put a template of stripe-public-key.js so users know where this file
should be and the format of the file.  This just makes things easier
when this code is downloaded for a new use/install.