Reorganize files to better structure per best practices on app engine site/documentation.
Place everything inside a "project" directory (stripe-appengine-frontend).  Then put the main file that loads the app (app.go) inside a "service" directory (process-cards) which is itself inside of a directory to house all "services" (services).  Then put all imported files (secondary files, not the file that is the root of the app) inside the pkgs directory.  This is done to make sure imports work correctly on the dev server and for deployment.

Update .gitignore to handle the new directory structure.