Rework almost everything, clean up, implement more company settings.
This is a huge commit with a lot of "barf".
 - Reworked almost every file to clean them up, revise comments, revise layout of code, reorganize code, better golang formating and styling to code.
 - Implemented functionality to add fees to company settings.  Also allow editing of statement descriptor in company settings.  Fixing some issues/bugs.
 - Use official context package instead of appengine/context where I can.
 - Break out company settings to a separate package to organize code better.

A lot more work to do still in this rework.  Have to implement deduction of fees in reports (show total charges and charges less fees), align columns in reports table correctly, use the newest stripe api, use the newest bootstrap 3.x and jquery files.  Redo README and install instructions.  Get cron task that removes old cards to work correctly.