Set path to templates and static files via app.yaml.  This was when the app is deployed on a non-appengine environment the user can provide the path to the files.
Serve configuration options to all templates.  This allows us to serve a header when we are in "appengine-dev" deployment type so it is easily visible in GUI.  Also allows us to change whether static files are served via CDN or on our local domain (where app is deployed).  This required redoing the templates to support new format of data being passed into templates.