Fix bug that was causing an error to be returned upon charging a card manually.
Error was generated because datastore couldn't update a card entity's LastUsedTimestamp because we didn't provide the key.ID when getting the full key.  This is because we didn't have the ID set in the card struct.  We previously didn't need this data because we weren't updating the card upon charging.  We need to get this data when getting card data which required getting the key and pulling out the ID value.  This issue would not have occured for SQLite because the ID column is retrieved upon a query (cloud datastore keeps track of ID in key, not is data of the entity).

Also, make sure the if an error does occur while updating the LastUsedTimestamp that we ignore it and only log it.