Implement Idempotency Key to prevent double charges from occuring.  This was added to solve issues with a downstream API requester sending charges to this app encountering an issue and retrying charges over and over resulting in duplicates.
Also implement go modules to handle dependencies better.  Upgraded to newest Stripe golang library.  Finally, implemented go-staticcheck since it is basically built into VS code now and it catches some small issues.

This is a relatively minor update in features and bugs fixed (idempotency key only), but touched a lot of files due to changes in imports (using versioned Stripe library) and go-staticcheck.  This project/repo/app is still maintained, just on a very limited basis since it is deemed feature complete at this point.  The code, organization, and implementation of some stuff is a pretty big mess and could use a heavy rewrite and clean up.  I am also working on rewriting this app in whole to a do the cleanup, not use appengine, add different payment processors, improve the feature set, and other details although it is a ways away from being ready.